Trying to blend flow image with grayscale images

+++ loaders.py
@@ -238,7 +238,7 @@ class KITTIVODatasetPreTransformed(Dataset):
         np_img1 = cv2.cvtColor(img1.permute(1,2,0).numpy(), cv2.COLOR_RGB2GRAY)
         np_img2 = cv2.cvtColor(img2.permute(1,2,0).numpy(), cv2.COLOR_RGB2GRAY)
         flow = cv2.calcOpticalFlowFarneback(np_img1, np_img2, None, 0.5, 3, 15, 3, 5, 1.2, 0)
-        stacked_img = torch.cat((torch.from_numpy(np_img1), torch.from_numpy(np_img2), torch.from_numpy(flow).permute(2,0,1)), 0)
+        stacked_img = torch.cat((torch.from_numpy(np_img1).float(), torch.from_numpy(np_img2).float(), torch.from_numpy(flow).permute(2,0,1)), 0)
         return stacked_img
